Label component rewritten to TS
