+++ src/components/Label/Label.mdx
@@ -1,5 +1,5 @@
 import { Playground, Props } from 'docz'
-import Label from '@components/Label'
+import { Label } from '@components/Label'
 import { Spacer } from '@components/Spacer'
 
 # Label
