+++ src/components/Label/Label.test.tsx
@@ -1,8 +1,7 @@
 import '@testing-library/jest-dom'
 
-import Label from '@components/Label'
+import { Label } from '@components/Label'
 import { render, screen } from '@testing-library/react'
-import React from 'react'
 
 describe('<Label/> mount', () => {
   const props = {
