--- src/components/Label/index.js
@@ -1,1 +1,0 @@
-export { default } from './Label'
